requirements.txt
discord.py==1.5.0
psycopg2==2.8.6
pandas==1.1.5
requests==2.25.1
https://download.pytorch.org/whl/cpu/torch-1.0.0-cp37-cp37m-linux_x86_64.whl
torchvision==0.9.1
Pillow==8.1.2